+++ pkg/heartbeat/types/types.go
@@ -1,12 +1,7 @@
 package types
 
 import (
-	"encoding/base64"
-	"encoding/json"
-
-	"github.com/pkg/errors"
 	appstatetypes "github.com/replicatedhq/replicated-sdk/pkg/appstate/types"
-	"github.com/replicatedhq/replicated-sdk/pkg/util"
 )
 
 type Distribution int64
@@ -89,36 +84,4 @@ type InstanceReportEvent struct {
 	DownstreamChannelID       string `json:"downstream_channel_id,omitempty"`
 	DownstreamChannelSequence int64  `json:"downstream_channel_sequence"`
 	DownstreamChannelName     string `json:"downstream_channel_name,omitempty"`
-}
-
-func (r *InstanceReport) Encode() ([]byte, error) {
-	data, err := json.Marshal(r)
-	if err != nil {
-		return nil, errors.Wrap(err, "failed to marshal instance report")
-	}
-	compressedData, err := util.GzipData(data)
-	if err != nil {
-		return nil, errors.Wrap(err, "failed to gzip instance report")
-	}
-	encodedData := base64.StdEncoding.EncodeToString(compressedData)
-
-	return []byte(encodedData), nil
-}
-
-func DecodeInstanceReport(encodedData []byte) (*InstanceReport, error) {
-	decodedData, err := base64.StdEncoding.DecodeString(string(encodedData))
-	if err != nil {
-		return nil, errors.Wrap(err, "failed to decode instance report")
-	}
-	decompressedData, err := util.GunzipData(decodedData)
-	if err != nil {
-		return nil, errors.Wrap(err, "failed to gunzip instance report")
-	}
-
-	instanceReport := InstanceReport{}
-	if err := json.Unmarshal(decompressedData, &instanceReport); err != nil {
-		return nil, errors.Wrap(err, "failed to unmarshal instance report")
-	}
-
-	return &instanceReport, nil
 }
