+++ pkg/store/memory_store.go
@@ -1,25 +1,11 @@
 package store
 
 import (
-	"context"
-
-	"github.com/pkg/errors"
 	kotsv1beta1 "github.com/replicatedhq/kotskinds/apis/kots/v1beta1"
 	appstatetypes "github.com/replicatedhq/replicated-sdk/pkg/appstate/types"
-	heartbeattypes "github.com/replicatedhq/replicated-sdk/pkg/heartbeat/types"
 	sdklicensetypes "github.com/replicatedhq/replicated-sdk/pkg/license/types"
 	upstreamtypes "github.com/replicatedhq/replicated-sdk/pkg/upstream/types"
-	"github.com/replicatedhq/replicated-sdk/pkg/util"
-	corev1 "k8s.io/api/core/v1"
-	kuberneteserrors "k8s.io/apimachinery/pkg/api/errors"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/kubernetes"
-)
-
-const (
-	InstanceReportSecretName = "replicated-instance-report"
-	InstanceReportSecretKey  = "report"
-	InstanceReportEventLimit = 4000
 )
 
 type InMemoryStore struct {
@@ -178,84 +164,4 @@ func (s *InMemoryStore) GetUpdates() []upstreamtypes.ChannelRelease {
 
 func (s *InMemoryStore) SetUpdates(updates []upstreamtypes.ChannelRelease) {
 	s.updates = updates
-}
-
-func (s *InMemoryStore) CreateInstanceReportEvent(event heartbeattypes.InstanceReportEvent) error {
-	existingSecret, err := s.clientset.CoreV1().Secrets(s.GetNamespace()).Get(context.TODO(), InstanceReportSecretName, metav1.GetOptions{})
-	if err != nil && !kuberneteserrors.IsNotFound(err) {
-		return errors.Wrap(err, "failed to get airgap instance report secret")
-	} else if kuberneteserrors.IsNotFound(err) {
-		instanceReport := &heartbeattypes.InstanceReport{
-			Events: []heartbeattypes.InstanceReportEvent{event},
-		}
-		data, err := instanceReport.Encode()
-		if err != nil {
-			return errors.Wrap(err, "failed to encode instance report")
-		}
-
-		uid, err := util.GetReplicatedDeploymentUID(s.clientset, s.GetNamespace())
-		if err != nil {
-			return errors.Wrap(err, "failed to get replicated deployment uid")
-		}
-
-		secret := &corev1.Secret{
-			TypeMeta: metav1.TypeMeta{
-				APIVersion: "v1",
-				Kind:       "ConfigMap",
-			},
-			ObjectMeta: metav1.ObjectMeta{
-				Name:      InstanceReportSecretName,
-				Namespace: s.GetNamespace(),
-				OwnerReferences: []metav1.OwnerReference{
-					{
-						APIVersion: "apps/v1",
-						Kind:       "Deployment",
-						Name:       util.GetReplicatedDeploymentName(),
-						UID:        uid,
-					},
-				},
-			},
-			Data: map[string][]byte{
-				InstanceReportSecretKey: data,
-			},
-		}
-
-		_, err = s.clientset.CoreV1().Secrets(s.GetNamespace()).Create(context.TODO(), secret, metav1.CreateOptions{})
-		if err != nil {
-			return errors.Wrap(err, "failed to create airgap instance report secret")
-		}
-
-		return nil
-	}
-
-	if existingSecret.Data == nil {
-		existingSecret.Data = map[string][]byte{}
-	}
-
-	existingInstanceReport := &heartbeattypes.InstanceReport{}
-	if existingSecret.Data[InstanceReportSecretKey] != nil {
-		existingInstanceReport, err = heartbeattypes.DecodeInstanceReport(existingSecret.Data[InstanceReportSecretKey])
-		if err != nil {
-			return errors.Wrap(err, "failed to load existing instance report")
-		}
-	}
-
-	existingInstanceReport.Events = append(existingInstanceReport.Events, event)
-	if len(existingInstanceReport.Events) > InstanceReportEventLimit {
-		existingInstanceReport.Events = existingInstanceReport.Events[len(existingInstanceReport.Events)-InstanceReportEventLimit:]
-	}
-
-	data, err := existingInstanceReport.Encode()
-	if err != nil {
-		return errors.Wrap(err, "failed to encode existing instance report")
-	}
-
-	existingSecret.Data[InstanceReportSecretKey] = data
-
-	_, err = s.clientset.CoreV1().Secrets(s.GetNamespace()).Update(context.TODO(), existingSecret, metav1.UpdateOptions{})
-	if err != nil {
-		return errors.Wrap(err, "failed to update airgap instance report secret")
-	}
-
-	return nil
 }
