+++ pkg/store/mock/mock_store.go
@@ -10,9 +10,8 @@ import (
 	gomock "github.com/golang/mock/gomock"
 	v1beta1 "github.com/replicatedhq/kotskinds/apis/kots/v1beta1"
 	types "github.com/replicatedhq/replicated-sdk/pkg/appstate/types"
-	types0 "github.com/replicatedhq/replicated-sdk/pkg/heartbeat/types"
-	types1 "github.com/replicatedhq/replicated-sdk/pkg/license/types"
-	types2 "github.com/replicatedhq/replicated-sdk/pkg/upstream/types"
+	types0 "github.com/replicatedhq/replicated-sdk/pkg/license/types"
+	types1 "github.com/replicatedhq/replicated-sdk/pkg/upstream/types"
 )
 
 // MockStore is a mock of Store interface.
@@ -36,20 +35,6 @@ func NewMockStore(ctrl *gomock.Controller) *MockStore {
 // EXPECT returns an object that allows the caller to indicate expected use.
 func (m *MockStore) EXPECT() *MockStoreMockRecorder {
 	return m.recorder
-}
-
-// CreateInstanceReportEvent mocks base method.
-func (m *MockStore) CreateInstanceReportEvent(event types0.InstanceReportEvent) error {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "CreateInstanceReportEvent", event)
-	ret0, _ := ret[0].(error)
-	return ret0
-}
-
-// CreateInstanceReportEvent indicates an expected call of CreateInstanceReportEvent.
-func (mr *MockStoreMockRecorder) CreateInstanceReportEvent(event interface{}) *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "CreateInstanceReportEvent", reflect.TypeOf((*MockStore)(nil).CreateInstanceReportEvent), event)
 }
 
 // GetAppID mocks base method.
@@ -165,10 +150,10 @@ func (mr *MockStoreMockRecorder) GetLicense() *gomock.Call {
 }
 
 // GetLicenseFields mocks base method.
-func (m *MockStore) GetLicenseFields() types1.LicenseFields {
+func (m *MockStore) GetLicenseFields() types0.LicenseFields {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "GetLicenseFields")
-	ret0, _ := ret[0].(types1.LicenseFields)
+	ret0, _ := ret[0].(types0.LicenseFields)
 	return ret0
 }
 
@@ -263,10 +248,10 @@ func (mr *MockStoreMockRecorder) GetReplicatedID() *gomock.Call {
 }
 
 // GetUpdates mocks base method.
-func (m *MockStore) GetUpdates() []types2.ChannelRelease {
+func (m *MockStore) GetUpdates() []types1.ChannelRelease {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "GetUpdates")
-	ret0, _ := ret[0].([]types2.ChannelRelease)
+	ret0, _ := ret[0].([]types1.ChannelRelease)
 	return ret0
 }
 
@@ -329,7 +314,7 @@ func (mr *MockStoreMockRecorder) SetLicense(license interface{}) *gomock.Call {
 }
 
 // SetLicenseFields mocks base method.
-func (m *MockStore) SetLicenseFields(licenseFields types1.LicenseFields) {
+func (m *MockStore) SetLicenseFields(licenseFields types0.LicenseFields) {
 	m.ctrl.T.Helper()
 	m.ctrl.Call(m, "SetLicenseFields", licenseFields)
 }
@@ -341,7 +326,7 @@ func (mr *MockStoreMockRecorder) SetLicenseFields(licenseFields interface{}) *go
 }
 
 // SetUpdates mocks base method.
-func (m *MockStore) SetUpdates(updates []types2.ChannelRelease) {
+func (m *MockStore) SetUpdates(updates []types1.ChannelRelease) {
 	m.ctrl.T.Helper()
 	m.ctrl.Call(m, "SetUpdates", updates)
 }
