+++ pkg/util/compress.go
@@ -4,6 +4,8 @@ import (
 	"bytes"
 	"compress/gzip"
 	"io"
+
+	"github.com/pkg/errors"
 )
 
 func GzipData(input []byte) ([]byte, error) {
@@ -12,12 +14,12 @@ func GzipData(input []byte) ([]byte, error) {
 
 	_, err := gw.Write(input)
 	if err != nil {
-		return nil, err
+		return nil, errors.Wrap(err, "failed to write to gzip writer")
 	}
 
 	err = gw.Close()
 	if err != nil {
-		return nil, err
+		return nil, errors.Wrap(err, "failed to close gzip writer")
 	}
 
 	return buf.Bytes(), nil
@@ -27,13 +29,13 @@ func GunzipData(input []byte) ([]byte, error) {
 	r := bytes.NewReader(input)
 	gr, err := gzip.NewReader(r)
 	if err != nil {
-		return nil, err
+		return nil, errors.Wrap(err, "failed to create gzip reader")
 	}
 	defer gr.Close()
 
 	decompressedData, err := io.ReadAll(gr)
 	if err != nil {
-		return nil, err
+		return nil, errors.Wrap(err, "failed to read from gzip reader")
 	}
 
 	return decompressedData, nil
